I-#75: Moved aws-amplify actions to common directory
