+++ clients/commons/config/amplify.ts
@@ -1,0 +1,6 @@
+import { Amplify } from 'aws-amplify';
+import awsExports from './awsExports';
+
+Amplify.configure(awsExports);
+
+export default Amplify;
