+++ clients/commons/hooks/useAuthenticatedUser.ts
